reorganize wasm unit testing
