+++ tests/wasm/testdata/original.mjs
@@ -1,0 +1,9 @@
+export class Up {
+	constructor() {
+		console.log('constructor')
+	}
+
+	fetch() {
+		console.log('fetch')
+	}
+}
